+++ petrify/src/test/java/com/github/exabrial/petrify/PetrifyTest.java
@@ -27,16 +27,16 @@ class PetrifyTest {
 		grove.nodesNodeIds = new int[] { 0, 1, 2 };
 		grove.nodesModes = new byte[] { 1, 0, 0 };
 		grove.nodesFeatureIds = new int[] { 0, 0, 0 };
-		grove.nodesValues = new float[] { 4.0f, 0.0f, 0.0f };
+		grove.nodesValues = new double[] { 4.0, 0.0, 0.0 };
 		grove.nodesTrueNodeIds = new int[] { 1, 0, 0 };
 		grove.nodesFalseNodeIds = new int[] { 2, 0, 0 };
-		grove.nodesHitRates = new float[] { 1.0f, 1.0f, 1.0f };
+		grove.nodesHitRates = new double[] { 1.0, 1.0, 1.0 };
 		grove.nodesMissingValueTracksTrue = new int[] { 0, 0, 0 };
 
 		grove.classTreeIds = new int[] { 0, 0 };
 		grove.classNodeIds = new int[] { 1, 2 };
 		grove.classIds = new int[] { 0, 1 };
-		grove.classWeights = new float[] { 1.0f, 1.0f };
+		grove.classWeights = new double[] { 1.0, 1.0 };
 
 		grove.classLabelsInt64s = new long[] { 0L, 1L };
 		grove.postTransform = (byte) 0;
